# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto MRP Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OkResponse {
  ok: boolean;
  message?: string;
}

export interface ErrorResponse {
  ok: boolean;
  error: string;
}

export interface WhatsappStatus {
  connected: boolean;
  status: string;
  number?: string | null;
}

export interface QrResponse {
  qr: string;
}

export interface WhitelistContact {
  id: number;
  phone: string;
  name: string;
  active: boolean;
  createdAt: string;
}

export interface CreateWhitelistContact {
  phone: string;
  name: string;
}

export interface UpdateWhitelistContact {
  active?: boolean;
  name?: string;
}

export interface HistoryEntry {
  id: number;
  phone: string;
  type: string;
  query: string;
  found: boolean;
  response: string;
  createdAt: string;
}

export interface SystemConfig {
  autoRespond: boolean;
  respondToGroups: boolean;
  maxDailyMessages: number;
  welcomeMessage: string;
}

export type ListHistoryParams = {
  limit?: number;
  phone?: string;
};