# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ZapAuto MRP Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whitelist:
    get:
      operationId: listWhitelist
      tags: [whitelist]
      summary: List all whitelist contacts
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhitelistContact"
    post:
      operationId: addWhitelistContact
      tags: [whitelist]
      summary: Add contact to whitelist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWhitelistContact"
      responses:
        "200":
          description: Contact added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistContact"
  /whitelist/{id}:
    delete:
      operationId: deleteWhitelistContact
      tags: [whitelist]
      summary: Remove contact from whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
    patch:
      operationId: updateWhitelistContact
      tags: [whitelist]
      summary: Toggle active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWhitelistContact"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistContact"
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
      required:
        - ok
...
    WhitelistContact:
      type: object
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - phone
        - name
        - active
        - createdAt
    CreateWhitelistContact:
      type: object
      properties:
        phone:
          type: string
        name:
          type: string
      required:
        - phone
        - name
    UpdateWhitelistContact:
      type: object
      properties:
        active:
          type: boolean
        name:
          type: string
...