# lib/api-zod/src/generated/types/createWhitelistContact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto MRP Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWhitelistContact {
  phone: string;
  name: string;
}