# lib/api-zod/src/generated/types/updateWhitelistContact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto MRP Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWhitelistContact {
  active?: boolean;
  name?: string;
}