# lib/api-zod/src/generated/types/whitelistContact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZapAuto MRP Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface WhitelistContact {
  id: number;
  phone: string;
  name: string;
  active: boolean;
  createdAt: Date;
}